# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateTripRequest:
    properties:
      available_seats:
        minimum: 1
        type: integer
      destination:
        type: string
      origin:
        type: string
      trip_date:
        type: string
      vehicle_id:
        type: integer
    required:
    - available_seats
    - destination
    - origin
    - trip_date
    - vehicle_id
    type: object
  models.CreateVehicleRequest:
    properties:
      seats_total:
        maximum: 8
        minimum: 1
        type: integer
    required:
    - seats_total
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.SearchTripResponse:
    properties:
      is_first_in_queue:
        type: boolean
      search_session_id:
        type: integer
      trip:
        $ref: '#/definitions/models.TripOffer'
      warning:
        type: string
    type: object
  models.SearchTripsRequest:
    properties:
      destination:
        type: string
      origin:
        type: string
      requested_seats:
        minimum: 1
        type: integer
      trip_date:
        type: string
    required:
    - destination
    - origin
    - requested_seats
    - trip_date
    type: object
...
  models.TripOffer:
    properties:
      available_seats:
        type: integer
      created_at:
        type: string
      destination:
        type: string
      driver_id:
        type: integer
      id:
        type: integer
      origin:
        type: string
      status:
        type: string
      total_seats:
        type: integer
      trip_date:
        type: string
      updated_at:
        type: string
      vehicle_id:
        type: integer
    type: object
...
  models.Vehicle:
    properties:
      created_at:
        type: string
      id:
        type: integer
      seats_total:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: MVP backend for Safar ride-sharing platform
  title: Safar Backend API
  version: "1.0"
paths:
...
  /driver/trips:
    get:
      description: Returns trip offers created by the authenticated driver.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TripOffer'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my trip offers
      tags:
      - Driver Trips
    post:
      consumes:
      - application/json
      description: Creates a driver trip offer using one of current user's vehicles.
      parameters:
      - description: Create trip request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTripRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TripOffer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create trip offer
      tags:
      - Driver Trips
...
  /passenger/search:
    post:
      consumes:
      - application/json
      description: Creates a search session and returns the first matching trip in
        queue.
      parameters:
      - description: Search trips request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SearchTripsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchTripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search trips queue
      tags:
      - Passenger Search
  /passenger/search/{sessionId}/next:
    post:
      description: Returns the next matching trip from an existing search session
        queue.
      parameters:
      - description: Search session ID
        in: path
        name: sessionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchTripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get next trip in queue
      tags:
      - Passenger Search
  /vehicles:
    get:
      description: Returns vehicles owned by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vehicle'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my vehicles
      tags:
      - Vehicles
    post:
      consumes:
      - application/json
      description: Creates a vehicle for the authenticated user.
      parameters:
      - description: Create vehicle request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vehicle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create vehicle
      tags:
      - Vehicles
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"